Refactor API types into modules
Since we'll soon be adding several new API endpoints to manage saved
UDFs, this first refactors the API objects.
