+++ arroyo-connectors/src/sse.rs
@@ -10,7 +10,7 @@ use futures::StreamExt;
 use tokio::sync::mpsc::Sender;
 use typify::import_types;
 
-use arroyo_rpc::types::{ConnectionSchema, ConnectionType, TestSourceMessage};
+use arroyo_rpc::api_types::connections::{ConnectionSchema, ConnectionType, TestSourceMessage};
 use serde::{Deserialize, Serialize};
 
 use crate::{pull_opt, Connection, EmptyConfig};
@@ -33,8 +33,8 @@ impl Connector for SSEConnector {
         "sse"
     }
 
-    fn metadata(&self) -> arroyo_rpc::types::Connector {
-        arroyo_rpc::types::Connector {
+    fn metadata(&self) -> arroyo_rpc::api_types::connections::Connector {
+        arroyo_rpc::api_types::connections::Connector {
             id: "sse".to_string(),
             name: "Server-Sent Events".to_string(),
             icon: ICON.to_string(),
