--- arroyo-rpc/src/types.rs
@@ -1,554 +1,0 @@
-use crate::formats::{Format, Framing};
-use crate::grpc as grpc_proto;
-use crate::grpc::api as api_proto;
-use anyhow::bail;
-use serde::{Deserialize, Serialize};
-use std::fmt::{Display, Formatter};
-use utoipa::{IntoParams, ToSchema};
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidateQueryPost {
-    pub query: String,
-    pub udfs: Option<Vec<Udf>>, // needed for query validation but are not themselves validated
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryValidationResult {
-    pub graph: Option<PipelineGraph>,
-    pub errors: Option<Vec<String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ValidateUdfsPost {
-    pub udfs_rs: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct UdfValidationResult {
-    pub udfs_rs: Option<String>,
-    pub errors: Option<Vec<String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelinePost {
-    pub name: String,
-    pub query: String,
-    pub udfs: Option<Vec<Udf>>,
-    pub preview: Option<bool>,
-    pub parallelism: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelinePatch {
-    pub parallelism: Option<u64>,
-    pub checkpoint_interval_micros: Option<u64>,
-    pub stop: Option<StopType>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelineRestart {
-    pub force: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Pipeline {
-    pub id: String,
-    pub name: String,
-    pub query: String,
-    pub udfs: Vec<Udf>,
-    pub checkpoint_interval_micros: u64,
-    pub stop: StopType,
-    pub created_at: u64,
-    pub action: Option<StopType>,
-    pub action_text: String,
-    pub action_in_progress: bool,
-    pub graph: PipelineGraph,
-    pub preview: bool,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelineGraph {
-    pub nodes: Vec<PipelineNode>,
-    pub edges: Vec<PipelineEdge>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelineNode {
-    pub node_id: String,
-    pub operator: String,
-    pub parallelism: u32,
-}
-
-impl From<api_proto::JobNode> for PipelineNode {
-    fn from(value: api_proto::JobNode) -> Self {
-        PipelineNode {
-            node_id: value.node_id,
-            operator: value.operator,
-            parallelism: value.parallelism,
-        }
-    }
-}
-
-impl From<api_proto::JobEdge> for PipelineEdge {
-    fn from(value: api_proto::JobEdge) -> Self {
-        PipelineEdge {
-            src_id: value.src_id,
-            dest_id: value.dest_id,
-            key_type: value.key_type,
-            value_type: value.value_type,
-            edge_type: value.edge_type,
-        }
-    }
-}
-
-impl From<api_proto::JobGraph> for PipelineGraph {
-    fn from(value: api_proto::JobGraph) -> Self {
-        PipelineGraph {
-            nodes: value.nodes.into_iter().map(|n| n.into()).collect(),
-            edges: value.edges.into_iter().map(|n| n.into()).collect(),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct PipelineEdge {
-    pub src_id: String,
-    pub dest_id: String,
-    pub key_type: String,
-    pub value_type: String,
-    pub edge_type: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub enum StopType {
-    None,
-    Checkpoint,
-    Graceful,
-    Immediate,
-    Force,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Job {
-    pub id: String,
-    pub running_desired: bool,
-    pub state: String,
-    pub run_id: u64,
-    pub start_time: Option<u64>,
-    pub finish_time: Option<u64>,
-    pub tasks: Option<u64>,
-    pub failure_message: Option<String>,
-    pub created_at: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum UdfLanguage {
-    Rust,
-}
-
-impl From<api_proto::UdfLanguage> for UdfLanguage {
-    fn from(value: api_proto::UdfLanguage) -> Self {
-        match value {
-            api_proto::UdfLanguage::Rust => UdfLanguage::Rust,
-        }
-    }
-}
-
-impl From<api_proto::Udf> for Udf {
-    fn from(value: api_proto::Udf) -> Self {
-        Udf {
-            language: value.language().into(),
-            definition: value.definition,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Udf {
-    pub language: UdfLanguage,
-    pub definition: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub enum JobLogLevel {
-    Info,
-    Warn,
-    Error,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct JobLogMessage {
-    pub id: String,
-    pub created_at: u64,
-    pub operator_id: Option<String>,
-    pub task_index: Option<u64>,
-    pub level: JobLogLevel,
-    pub message: String,
-    pub details: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Checkpoint {
-    pub epoch: u32,
-    pub backend: String,
-    pub start_time: u64,
-    pub finish_time: Option<u64>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum CheckpointSpanType {
-    Alignment,
-    Sync,
-    Async,
-    Committing,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct CheckpointEventSpan {
-    pub start_time: u64,
-    pub finish_time: u64,
-    pub span_type: CheckpointSpanType,
-    pub description: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SubtaskCheckpointGroup {
-    pub index: u32,
-    pub bytes: u64,
-    pub event_spans: Vec<CheckpointEventSpan>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct OperatorCheckpointGroup {
-    pub operator_id: String,
-    pub bytes: u64,
-    pub subtasks: Vec<SubtaskCheckpointGroup>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-#[aliases(
-    PipelineCollection = PaginatedCollection<Pipeline>,
-    JobLogMessageCollection = PaginatedCollection<JobLogMessage>,
-    ConnectionTableCollection = PaginatedCollection<ConnectionTable>,
-)]
-pub struct PaginatedCollection<T> {
-    pub data: Vec<T>,
-    pub has_more: bool,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-#[aliases(
-    JobCollection = NonPaginatedCollection<Job>,
-    OperatorCheckpointGroupCollection = NonPaginatedCollection<OperatorCheckpointGroup>,
-    CheckpointCollection = NonPaginatedCollection<Checkpoint>,
-    OperatorMetricGroupCollection = NonPaginatedCollection<OperatorMetricGroup>,
-    ConnectorCollection = NonPaginatedCollection<Connector>,
-    ConnectionProfileCollection = NonPaginatedCollection<ConnectionProfile>,
-)]
-pub struct NonPaginatedCollection<T> {
-    pub data: Vec<T>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct OutputData {
-    pub operator_id: String,
-    pub timestamp: u64,
-    pub key: String,
-    pub value: String,
-}
-
-impl From<grpc_proto::OutputData> for OutputData {
-    fn from(value: grpc_proto::OutputData) -> Self {
-        OutputData {
-            operator_id: value.operator_id,
-            timestamp: value.timestamp,
-            key: value.key,
-            value: value.value,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, Hash, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum MetricNames {
-    BytesRecv,
-    BytesSent,
-    MessagesRecv,
-    MessagesSent,
-    Backpressure,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Metric {
-    pub time: u64,
-    pub value: f64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SubtaskMetrics {
-    pub index: u32,
-    pub metrics: Vec<Metric>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct MetricGroup {
-    pub name: MetricNames,
-    pub subtasks: Vec<SubtaskMetrics>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct OperatorMetricGroup {
-    pub operator_id: String,
-    pub metric_groups: Vec<MetricGroup>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct Connector {
-    pub id: String,
-    pub name: String,
-    pub icon: String,
-    pub description: String,
-    pub table_config: String,
-    pub enabled: bool,
-    pub source: bool,
-    pub sink: bool,
-    pub custom_schemas: bool,
-    pub testing: bool,
-    pub hidden: bool,
-    pub connection_config: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionProfile {
-    pub id: String,
-    pub name: String,
-    pub connector: String,
-    pub config: serde_json::Value,
-    pub description: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionProfilePost {
-    pub name: String,
-    pub connector: String,
-    pub config: serde_json::Value,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum ConnectionType {
-    Source,
-    Sink,
-}
-
-impl Display for ConnectionType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ConnectionType::Source => write!(f, "SOURCE"),
-            ConnectionType::Sink => write!(f, "SINK"),
-        }
-    }
-}
-
-impl TryFrom<String> for ConnectionType {
-    type Error = String;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        match value.to_lowercase().as_str() {
-            "source" => Ok(ConnectionType::Source),
-            "sink" => Ok(ConnectionType::Sink),
-            _ => Err(format!("Invalid connection type: {}", value)),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum PrimitiveType {
-    Int32,
-    Int64,
-    UInt32,
-    UInt64,
-    F32,
-    F64,
-    Bool,
-    String,
-    Bytes,
-    UnixMillis,
-    UnixMicros,
-    UnixNanos,
-    DateTime,
-    Json,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct StructType {
-    pub name: Option<String>,
-    pub fields: Vec<SourceField>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum FieldType {
-    Primitive(PrimitiveType),
-    Struct(StructType),
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct SourceFieldType {
-    pub r#type: FieldType,
-    pub sql_name: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct SourceField {
-    pub field_name: String,
-    pub field_type: SourceFieldType,
-    pub nullable: bool,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum SchemaDefinition {
-    JsonSchema(String),
-    ProtobufSchema(String),
-    AvroSchema(String),
-    RawSchema(String),
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionSchema {
-    pub format: Option<Format>,
-    pub framing: Option<Framing>,
-    pub struct_name: Option<String>,
-    pub fields: Vec<SourceField>,
-    pub definition: Option<SchemaDefinition>,
-}
-
-impl ConnectionSchema {
-    pub fn try_new(
-        format: Option<Format>,
-        framing: Option<Framing>,
-        struct_name: Option<String>,
-        fields: Vec<SourceField>,
-        definition: Option<SchemaDefinition>,
-    ) -> anyhow::Result<Self> {
-        let s = ConnectionSchema {
-            format,
-            framing,
-            struct_name,
-            fields,
-            definition,
-        };
-
-        s.validate()
-    }
-
-    pub fn validate(self) -> anyhow::Result<Self> {
-        match &self.format {
-            Some(Format::RawString(_)) => {
-                if self.fields.len() != 1
-                    || self.fields.get(0).unwrap().field_type.r#type
-                        != FieldType::Primitive(PrimitiveType::String)
-                {
-                    bail!("raw_string format requires a schema with a single field of type TEXT");
-                }
-            }
-            _ => {}
-        }
-
-        Ok(self)
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema, IntoParams)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionTable {
-    #[serde(skip_serializing)]
-    pub id: i64,
-    #[serde(rename = "id")]
-    pub pub_id: String,
-    pub name: String,
-    pub created_at: u64,
-    pub connector: String,
-    pub connection_profile: Option<ConnectionProfile>,
-    pub table_type: ConnectionType,
-    pub config: serde_json::Value,
-    pub schema: ConnectionSchema,
-    pub consumers: u32,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ConnectionTablePost {
-    pub name: String,
-    pub connector: String,
-    pub connection_profile_id: Option<String>,
-    pub config: serde_json::Value,
-    pub schema: Option<ConnectionSchema>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct TestSourceMessage {
-    pub error: bool,
-    pub done: bool,
-    pub message: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ConfluentSchema {
-    pub schema: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, IntoParams)]
-#[serde(rename_all = "camelCase")]
-pub struct ConfluentSchemaQueryParams {
-    pub endpoint: String,
-    pub topic: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, IntoParams, ToSchema)]
-#[into_params(parameter_in = Query)]
-#[serde(rename_all = "snake_case")]
-pub struct PaginationQueryParams {
-    pub starting_after: Option<String>,
-    pub limit: Option<u32>,
-}
